Add SugarCube header and make it the default.
Added the SugarCube header from http://www.motoslave.net/sugarcube/, which is
an enhanced fork of the Sugarcane header included with stock twine. There have
been a number of fixes to various functions as well as additional macros that
make authoring twine stories easier. Used with permission of the author.