Fix toRSS function as well as add SugarCube specific link handling
SugarCube has a number of additional link types to passages that need to be
supported in the link handler. These include bind, link, back, and return
macros. Also the toRSS functionality of the library was fixed, though will
likely be removed at a later date, due to being unused within any of the twine
code.