Add media tag with automatic base64 encoding of image files
Added a media tag which enables any passages with the media tag to be parsed
for the filenames of image files, which will then be automatically loaded,
base64 encoded, and embedded within the output html file. This basically means
that it is quite simple to have images contained within stories but still
keeping the simplicity of a single html story file containing the entire story.
There are some restrictions to keep in mind though as IE doesn't have very
proper support, only being supported in IE8 and above, and in that version,
having a 32k file size limit for the encoded data. Later IE versions do not
have this limitation.