Add font (.woff) support to base64 media encoding
Add encoding support for the .woff font format. This was chosen as it seems to
be the most compatible between browsers, without adding many different font
types that will just make the output file larger with minimal gains. This is
put into a for loop in case in the future further font types are needed.