Look for Media files in source directory rather than output directory
Rather than looking in the output directory for the files to base64 encode,
look within the source directory. Uses the same sort of code that the
StoryIncludes uses, which necessitated bringing the save directory down the
function chain to the encode text function. Not the cleanest solution for the
problem, but the only one that seemed to be feasible without changing a large
amount of code. Also added in the tags into the encode text function
declaration, as it was forgotten.